Add files via upload
This is my attempt at health system which ultimately was not working as intended. This was 2nd version of the game since start. I had packed it into 3 part file as github does not allow files bigger than 25mb so please use Winrar to unpack in case of problems and download main zip file as well as the two .z0x files.